+++ test/resources/language/haskell/stack-project-with-renamed-stack-yaml/src/Foo.hs
@@ -1,0 +1,5 @@
+module Foo (foo) where
+
+foo :: a -> a
+foo x = x
+